consolidated TaskScenario extensions to single modules
